requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
numpy>=1.24.3
matplotlib>=3.7.1
seaborn>=0.12.2
gensim>=4.3.1
nltk>=3.8.1
scikit-learn>=1.2.2
sentence-transformers>=2.2.2
textstat>=0.7.3
requests>=2.31.0
python-dotenv>=1.0.0
langdetect>=1.0.9
openpyxl>=3.1.2
